# Speed up SSL_add_{file,dir}_cert_subjects_to_stack