@neverpanic 2024-07-31
I sampled an stunnel process with `perf` during startup loading a large number of certificates to identify that the allocation in the comparison function is root cause for the slowness. This was exacerbated by stunnel's use of a custom allocator, which makes `OPENSSL_malloc()` even slower and makes this issue much more severe. However, this is also an issue with the default allocator with large numbers of certificates, as evident from the benchmark data.