@neverpanic 2024-07-31
Flame graph of an stunnel startup with 1000 certificates in a `CApath` before the change:
![before](https://github.com/user-attachments/assets/7146d34e-17c3-43b1-a282-7eebe6de8d35)

The same startup with `LD_LIBRARY_PATH` pointing to a copy of OpenSSL with this PR applied:
![after](https://github.com/user-attachments/assets/ac14515f-d726-485f-835d-815793bad8f6)